Big overhaul
Guys, I realized that there was a great solution to our big problem with how we're going to connect the controller and the classes and the user interface and what not. Now, the controller literally controls ALL the data. The "PeopleClasses" don't keep track of their own relationships anymore. The controller keeps three huge lists of people, relationship types, and relationships, and processes whatever the user wants and sends back the data. It's better to see it working in action than me telling you though. The program now runs (choose to ignore errors), but all you can do is view people and change the currentPerson. The big changes are:
- New class 'Relationships'
- Refined old methods and added new ones in Controller
- Some name changes in 'PeopleClass' and 'RelationType'